requirements.txt
Flask==3.0.2
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
gunicorn==21.2.0
bcrypt==4.1.3
mysql-connector-python==8.3.0
python-dotenv==1.0.1
flake8==7.0.0
safety==3.0.1
pip-audit==2.7.2